ci: Move several tests out of smoke test jobs
