+++ tests/integration/integration_go_smoke_test.go
@@ -1,4 +1,17 @@
-// Copyright 2016-2021, Pulumi Corporation.  All rights reserved.
+// Copyright 2016-2022, Pulumi Corporation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 //go:build go || all
 
 package ints
@@ -161,15 +174,4 @@ func optsForConstructGo(t *testing.T, dir string, expectedResourceCount int, loc
 			}
 		},
 	}
-}
-
-// TestRefreshGo simply tests that we can build and run an empty Go project with the `refresh` option set.
-func TestRefreshGo(t *testing.T) {
-	integration.ProgramTest(t, &integration.ProgramTestOptions{
-		Dir: filepath.Join("refresh", "go"),
-		Dependencies: []string{
-			"github.com/pulumi/pulumi/sdk/v3",
-		},
-		Quick: true,
-	})
 }
